remove old lexer
