+++ src/lexer/mod.rs
@@ -1,227 +1,251 @@
-pub mod rewrite;
-mod stack;
+pub mod stack;
 
 /// Defines the wrapper type for an error returned from Lexing
 type LexError = String;
 
 /// Lex returns either a
-pub type LexResult = Result<Vec<Token>, LexError>;
+pub type LexResult<'a> = Result<Vec<Token<'a>>, LexError>;
+
+/// Provides metadata about the classification of a given Token.
+#[derive(Debug)]
+#[allow(unused)]
+pub enum TokenClass {
+    Identifier,
+    Keywords,
+    Constant,
+    StringLiteral,
+    Operator,
+    Separator,
+}
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum TokenType {
-    Eof,
+pub enum TokenKind {
+    Identifier,
 
-    // Binary operators
-    Equal,
-    EqualEqual,
-    Pipe,
-    PipePipe,
+    // Constants
+    FloatConstant,
+    CharacterConstant,
+    IntegerConstant,
+
+    StringLiteral,
+
+    // Keywords
+    Auto,
+    Break,
+    Case,
+    Char,
+    Const,
+    Continue,
+    Default,
+    Do,
+    Double,
+    Else,
+    Enum,
+    Extern,
+    Float,
+    For,
+    Goto,
+    If,
+    Int,
+    Long,
+    Register,
+    Return,
+    Short,
+    Signed,
+    SizeOf,
+    Static,
+    Struct,
+    Switch,
+    TypeDef,
+    Union,
+    Unsigned,
+    Void,
+    Volatile,
+    While,
+
+    // Operators
     Ampersand,
     AmpersandAmpersand,
-    Carat,
+    AmpersandEqual,
+    Arrow,
     Bang,
     BangEqual,
-    LessThan,
-    GreaterThan,
-    LessEqual,
-    GreaterEqual,
-    LShift,
-    RShift,
-    Plus,
-    Minus,
-    Star,
-    Slash,
-    PercentSign,
-
-    // Other operators
-    Tilde,
-    PlusPlus,
-    MinusMinus,
-
-    // Type keywords
-    Auto,
-    Signed,
-    Unsigned,
-    Void,
-    Char,
-    Short,
-    Int,
-    Long,
-
-    // Other keywords
-    If,
-    Else,
-    While,
-    For,
-    Return,
-
-    // Structural tokens
-    CharLiteral,
-    IntLiteral,
-    StrLiteral,
-    SemiColon,
-    Identifier,
-    LBrace,
-    RBrace,
-    LParen,
-    RParen,
-    LBracket,
-    RBracket,
+    Caret,
+    CaretEqual,
+    Colon,
     Comma,
-    Newline,
-    Space,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Token {
-    Eof,
-
-    // Binary operators
+    Dot,
     Equal,
     EqualEqual,
-    Pipe,
-    PipePipe,
-    Ampersand,
-    AmpersandAmperand,
-    Carat,
-    Bang,
-    BangEqual,
-    LessThan,
-    GreaterThan,
-    LessEqual,
     GreaterEqual,
-    LShift,
-    RShift,
-    Plus,
+    GreaterThan,
+    LeftBrace,
+    LeftBracket,
+    LeftParen,
+    LeftShift,
+    LeftShiftEqual,
+    LessEqual,
+    LessThan,
     Minus,
-    Star,
-    Slash,
-    PercentSign,
-
-    // Other operators
-    Tilde,
-    PlusPlus,
+    MinusEqual,
     MinusMinus,
+    PercentEqual,
+    PercentSign,
+    Pipe,
+    PipeEqual,
+    PipePipe,
+    Plus,
+    PlusEqual,
+    PlusPlus,
+    Question,
+    RightBrace,
+    RightBracket,
+    RightParen,
+    RightShift,
+    RightShiftEqual,
+    Slash,
+    SlashEqual,
+    Star,
+    StarEqual,
+    Tilde,
 
-    // Type keywords
-    Auto,
-    Signed,
-    Unsigned,
-    Void,
-    Char,
-    Short,
-    Int,
-    Long,
-
-    // Other keywords
-    If,
-    Else,
-    While,
-    For,
-    Return,
-
-    // Structural tokens
-    CharLiteral(char),
-    IntLiteral(isize),
-    StrLiteral(String),
-    Identifier(String),
+    // Separators
     SemiColon,
-    LBrace,
-    RBrace,
-    LParen,
-    RParen,
-    LBracket,
-    RBracket,
-    Comma,
-    Newline,
-    Space,
 }
 
-impl Token {
-    pub fn to_token_type(&self) -> TokenType {
+impl TokenKind {
+    pub fn token_class(&self) -> TokenClass {
         match self {
-            Token::Eof => TokenType::Eof,
-            Token::Equal => TokenType::Equal,
-            Token::Pipe => TokenType::Pipe,
-            Token::AmpersandAmperand => TokenType::AmpersandAmpersand,
-            Token::PipePipe => TokenType::PipePipe,
-            Token::Carat => TokenType::Carat,
-            Token::Ampersand => TokenType::Ampersand,
-            Token::EqualEqual => TokenType::EqualEqual,
-            Token::Bang => TokenType::Bang,
-            Token::BangEqual => TokenType::BangEqual,
-            Token::LessThan => TokenType::LessThan,
-            Token::GreaterThan => TokenType::GreaterThan,
-            Token::LessEqual => TokenType::LessEqual,
-            Token::GreaterEqual => TokenType::GreaterEqual,
-            Token::LShift => TokenType::LShift,
-            Token::RShift => TokenType::RShift,
-            Token::Plus => TokenType::Plus,
-            Token::Minus => TokenType::Minus,
-            Token::Star => TokenType::Star,
-            Token::Slash => TokenType::Slash,
-            Token::PercentSign => TokenType::PercentSign,
-            Token::Tilde => TokenType::Tilde,
-            Token::PlusPlus => TokenType::PlusPlus,
-            Token::MinusMinus => TokenType::MinusMinus,
-            Token::Auto => TokenType::Auto,
-            Token::Signed => TokenType::Signed,
-            Token::Unsigned => TokenType::Unsigned,
-            Token::Void => TokenType::Void,
-            Token::Char => TokenType::Char,
-            Token::Short => TokenType::Short,
-            Token::Int => TokenType::Int,
-            Token::Long => TokenType::Long,
-            Token::If => TokenType::If,
-            Token::Else => TokenType::Else,
-            Token::While => TokenType::While,
-            Token::For => TokenType::For,
-            Token::Return => TokenType::Return,
-            Token::CharLiteral(_) => TokenType::CharLiteral,
-            Token::IntLiteral(_) => TokenType::IntLiteral,
-            Token::StrLiteral(_) => TokenType::StrLiteral,
-            Token::Identifier(_) => TokenType::Identifier,
-            Token::SemiColon => TokenType::SemiColon,
-            Token::LBrace => TokenType::LBrace,
-            Token::RBrace => TokenType::RBrace,
-            Token::LParen => TokenType::LParen,
-            Token::RParen => TokenType::RParen,
-            Token::LBracket => TokenType::LBracket,
-            Token::RBracket => TokenType::RBracket,
-            Token::Comma => TokenType::Comma,
-            Token::Newline => TokenType::Newline,
-            Token::Space => TokenType::Space,
+            TokenKind::Auto
+            | TokenKind::Break
+            | TokenKind::Case
+            | TokenKind::Char
+            | TokenKind::Const
+            | TokenKind::Continue
+            | TokenKind::Default
+            | TokenKind::Do
+            | TokenKind::Double
+            | TokenKind::Else
+            | TokenKind::Enum
+            | TokenKind::Extern
+            | TokenKind::Float
+            | TokenKind::For
+            | TokenKind::Goto
+            | TokenKind::If
+            | TokenKind::Int
+            | TokenKind::Long
+            | TokenKind::Register
+            | TokenKind::Return
+            | TokenKind::Short
+            | TokenKind::Signed
+            | TokenKind::SizeOf
+            | TokenKind::Static
+            | TokenKind::Struct
+            | TokenKind::Switch
+            | TokenKind::TypeDef
+            | TokenKind::Union
+            | TokenKind::Unsigned
+            | TokenKind::Void
+            | TokenKind::Volatile
+            | TokenKind::While => TokenClass::Keywords,
+
+            TokenKind::FloatConstant
+            | TokenKind::IntegerConstant
+            | TokenKind::CharacterConstant => TokenClass::Constant,
+
+            TokenKind::StringLiteral => TokenClass::StringLiteral,
+
+            TokenKind::Identifier => TokenClass::Identifier,
+
+            TokenKind::Ampersand
+            | TokenKind::AmpersandAmpersand
+            | TokenKind::AmpersandEqual
+            | TokenKind::Arrow
+            | TokenKind::Bang
+            | TokenKind::BangEqual
+            | TokenKind::Caret
+            | TokenKind::CaretEqual
+            | TokenKind::Colon
+            | TokenKind::Comma
+            | TokenKind::Dot
+            | TokenKind::Equal
+            | TokenKind::EqualEqual
+            | TokenKind::GreaterEqual
+            | TokenKind::GreaterThan
+            | TokenKind::LeftBrace
+            | TokenKind::LeftBracket
+            | TokenKind::LeftParen
+            | TokenKind::LeftShift
+            | TokenKind::LeftShiftEqual
+            | TokenKind::LessEqual
+            | TokenKind::LessThan
+            | TokenKind::Minus
+            | TokenKind::MinusEqual
+            | TokenKind::MinusMinus
+            | TokenKind::PercentEqual
+            | TokenKind::PercentSign
+            | TokenKind::Pipe
+            | TokenKind::PipeEqual
+            | TokenKind::PipePipe
+            | TokenKind::Plus
+            | TokenKind::PlusEqual
+            | TokenKind::PlusPlus
+            | TokenKind::Question
+            | TokenKind::RightBrace
+            | TokenKind::RightBracket
+            | TokenKind::RightParen
+            | TokenKind::RightShift
+            | TokenKind::RightShiftEqual
+            | TokenKind::Slash
+            | TokenKind::SlashEqual
+            | TokenKind::Star
+            | TokenKind::StarEqual
+            | TokenKind::Tilde => TokenClass::Operator,
+
+            // Separator
+            TokenKind::SemiColon => TokenClass::Separator,
         }
     }
 }
 
-#[derive(Debug, Clone, Copy)]
-pub(crate) struct Cursor {
+/// Provides input tracking.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct Cursor {
     index: usize,
     col: usize,
     line: usize,
 }
 
 impl Cursor {
+    pub const fn new(index: usize, col: usize, line: usize) -> Self {
+        Self { index, col, line }
+    }
+
+    pub fn increment_column(mut self) -> Self {
+        self.increment_column_mut();
+        self
+    }
+
     pub fn increment_column_mut(&mut self) {
         self.index += 1;
         self.col += 1;
     }
 
+    pub fn increment_line(mut self) -> Self {
+        self.increment_line_mut();
+        self
+    }
+
     pub fn increment_line_mut(&mut self) {
         self.index += 1;
         self.line += 1;
-        self.col += 0;
+        self.col = 0;
     }
 }
 
 impl Default for Cursor {
     fn default() -> Self {
-        Self {
-            index: 0,
-            col: 0,
-            line: 1,
-        }
+        Self::new(0, 0, 1)
     }
 }
 
@@ -231,30 +255,87 @@ impl core::fmt::Display for Cursor {
     }
 }
 
-use std::iter::Iterator;
-
-const STACK_SIZE: usize = 4096;
-
-type LexerStack<const N: usize> = stack::Stack<TokenOrLexeme, N>;
-
-const TOKEN_OR_LEXEME_INITIALIZER: TokenOrLexeme = TokenOrLexeme::Lexeme('\x00');
-
-#[derive(Debug, Clone)]
-enum TokenOrLexeme {
-    Token(Token),
-    Lexeme(char),
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct Span {
+    start: Cursor,
+    end: Cursor,
 }
 
-impl TokenOrLexeme {
-    fn to_lexeme(&self) -> Option<char> {
-        match self {
-            TokenOrLexeme::Lexeme(c) => Some(*c),
-            TokenOrLexeme::Token(_) => None,
-        }
+impl Span {
+    pub fn new(start: Cursor, end: Cursor) -> Self {
+        Self { start, end }
     }
 }
 
-impl Default for TokenOrLexeme {
+impl From<Span> for std::ops::Range<usize> {
+    fn from(span: Span) -> Self {
+        let start = span.start.index;
+        let end = span.end.index;
+
+        start..end
+    }
+}
+
+impl core::fmt::Display for Span {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "({})..({})", self.start, self.end)
+    }
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct Token<'a> {
+    pub span: Span,
+    pub data: Option<&'a str>,
+    pub kind: TokenKind,
+}
+
+impl<'a> Token<'a> {
+    pub fn new(span: Span, kind: TokenKind) -> Self {
+        Self {
+            span,
+            data: None,
+            kind,
+        }
+    }
+
+    pub fn with_data(span: Span, data: &'a str, kind: TokenKind) -> Self {
+        Self {
+            span,
+            data: Some(data),
+            kind,
+        }
+    }
+
+    /// Returns the token's kind.
+    pub fn as_kind(&self) -> TokenKind {
+        self.kind
+    }
+
+    /// Returns the slice of a given input that a token represents. If the
+    /// token is out of range for an input, None is returned.
+    pub fn str_from_input(&self, input: &'a str) -> Option<&'a str> {
+        let range: std::ops::Range<_> = self.span.into();
+
+        input.get(range)
+    }
+}
+
+use std::iter::Iterator;
+
+const STACK_SIZE: usize = 2048;
+
+type LexerStack<'a, const N: usize> = stack::Stack<TokenOrLexeme<'a>, N>;
+
+const TOKEN_OR_LEXEME_INITIALIZER: TokenOrLexeme =
+    TokenOrLexeme::Lexeme(Cursor::new(0, 0, 1), '\x00');
+
+#[derive(Debug, Clone)]
+enum TokenOrLexeme<'a> {
+    Token(Token<'a>),
+    Lexeme(Cursor, char),
+}
+
+impl<'a> Default for TokenOrLexeme<'a> {
     fn default() -> Self {
         TOKEN_OR_LEXEME_INITIALIZER
     }
@@ -267,16 +348,16 @@ macro_rules! token {
 }
 
 macro_rules! lexeme {
-    ($lexeme:literal) => {
-        TokenOrLexeme::Lexeme($lexeme)
+    ($cursor:expr, $lexeme:literal) => {
+        TokenOrLexeme::Lexeme($cursor, $lexeme)
     };
-    ($lexeme:expr) => {
-        TokenOrLexeme::Lexeme($lexeme)
+    ($cursor:expr, $lexeme:expr) => {
+        TokenOrLexeme::Lexeme($cursor, $lexeme)
     };
 }
 
 #[derive(Debug, Clone, Copy)]
-enum EscapedAscii {
+pub enum EscapedAscii {
     /// Character was escaped
     Escaped(char),
     /// Represents a standard character
@@ -284,14 +365,18 @@ enum EscapedAscii {
 }
 
 impl EscapedAscii {
-    fn unwrap(self) -> char {
+    pub fn as_char(&self) -> char {
         match self {
-            Self::Escaped(c) | Self::NonEscaped(c) => c,
+            Self::Escaped(c) | Self::NonEscaped(c) => *c,
         }
+    }
+
+    pub fn unwrap(self) -> char {
+        self.as_char()
     }
 }
 
-fn to_ascii_escaped_char(c: char) -> EscapedAscii {
+pub fn to_ascii_escaped_char(c: char) -> EscapedAscii {
     use EscapedAscii::*;
 
     match c {
@@ -302,6 +387,39 @@ fn to_ascii_escaped_char(c: char) -> EscapedAscii {
     }
 }
 
+pub fn to_ascii_escaped_string<S: AsRef<str>>(input: S) -> String {
+    let mut iter = input.as_ref().chars().peekable();
+    let mut chars = vec![];
+    let mut curr = iter.next();
+    let mut next = iter.peek();
+    while curr.is_some() {
+        match (curr, next) {
+            (None, Some(_)) | (None, None) => return String::new(),
+            (Some(c), None) => chars.push(EscapedAscii::NonEscaped(c)),
+            (Some('\\'), Some(next_c)) => match to_ascii_escaped_char(*next_c) {
+                EscapedAscii::Escaped(c) => {
+                    chars.push(EscapedAscii::Escaped(c));
+                    // consume one extra char.
+                    iter.next();
+                }
+                EscapedAscii::NonEscaped(_) => chars.push(EscapedAscii::NonEscaped('\\')),
+            },
+            (Some(c), Some(_)) => chars.push(EscapedAscii::NonEscaped(c)),
+        }
+
+        curr = iter.next();
+        next = iter.peek();
+    }
+
+    chars.into_iter().map(|ea| ea.as_char()).collect::<String>()
+}
+
+#[derive(Debug)]
+enum AdvanceOperation {
+    IncrementLine,
+    IncrementColumn,
+}
+
 #[derive(Debug)]
 enum LexOperation<T> {
     // the last n elements to pop off and the target type
@@ -309,237 +427,964 @@ enum LexOperation<T> {
     ShiftReduce(usize, T),
     Reduce(usize, T),
     NoFurtherRefinement,
+    // usize represents an optional number of tokens to purge off the stack
+    // when advancing.
+    Advance(usize, AdvanceOperation),
     Err(String),
 }
 
 fn lexeme_is_identifier_ascii() -> impl Fn(&TokenOrLexeme) -> bool {
-    move |tol: &TokenOrLexeme| matches!(tol, TokenOrLexeme::Lexeme(l) if l.is_ascii_alphanumeric() || *l == '_')
+    move |tol: &TokenOrLexeme| matches!(tol, TokenOrLexeme::Lexeme(_, l) if l.is_ascii_alphanumeric() || *l == '_')
 }
 
 fn matches_rules<T>(src: &[T], pred: impl Fn(&T) -> bool) -> bool {
     src.iter().all(pred)
 }
 
-fn stack_eval_to_token(stack: &[TokenOrLexeme], next: Option<char>) -> LexOperation<TokenOrLexeme> {
+fn stack_eval_to_token<'a>(
+    source: &'a str,
+    stack: &[TokenOrLexeme],
+    next: Option<(Cursor, char)>,
+) -> LexOperation<TokenOrLexeme<'a>> {
     match (stack, next) {
         // shift an empty stack
-        ([], Some(c)) => LexOperation::Shift(lexeme!(c)),
+        ([], Some((cursor, ch))) => LexOperation::Shift(lexeme!(cursor, ch)),
 
-        ([lexeme!('{')], _) => LexOperation::Reduce(1, token!(Token::LBrace)),
-        ([lexeme!('}')], _) => LexOperation::Reduce(1, token!(Token::RBrace)),
-        ([lexeme!('(')], _) => LexOperation::Reduce(1, token!(Token::LParen)),
-        ([lexeme!(')')], _) => LexOperation::Reduce(1, token!(Token::RParen)),
-        ([lexeme!('[')], _) => LexOperation::Reduce(1, token!(Token::LBracket)),
-        ([lexeme!(']')], _) => LexOperation::Reduce(1, token!(Token::RBracket)),
-        ([lexeme!(';')], _) => LexOperation::Reduce(1, token!(Token::SemiColon)),
-        ([lexeme!(',')], _) => LexOperation::Reduce(1, token!(Token::Comma)),
+        ([TokenOrLexeme::Lexeme(cur, '{')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::LeftBrace
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '}')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::RightBrace
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '(')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::LeftParen
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, ')')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::RightParen
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '[')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::LeftBracket
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, ']')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::RightBracket
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, ';')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::SemiColon
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, ':')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::Colon
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, ',')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::Comma
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '.')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::Dot
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '?')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::Question
+            )),
+        ),
 
-        ([TokenOrLexeme::Token(Token::Equal)], Some('=')) => {
-            LexOperation::ShiftReduce(2, token!(Token::EqualEqual))
-        }
-        ([lexeme!('=')], _) => LexOperation::Reduce(1, token!(Token::Equal)),
-        ([TokenOrLexeme::Token(Token::Bang)], Some('=')) => {
-            LexOperation::ShiftReduce(2, token!(Token::BangEqual))
-        }
-        ([lexeme!('!')], _) => LexOperation::Reduce(1, token!(Token::Bang)),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::Minus,
+                ..
+            })],
+            Some((end, '>')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::Arrow,
+            )),
+        ),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::Equal,
+                ..
+            })],
+            Some((end, '=')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::EqualEqual,
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '=')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::Equal
+            )),
+        ),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::Bang,
+                ..
+            })],
+            Some((end, '=')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::BangEqual
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '!')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::Bang
+            )),
+        ),
 
-        ([TokenOrLexeme::Token(Token::GreaterThan)], Some('>')) => {
-            LexOperation::ShiftReduce(2, token!(Token::RShift))
-        }
-        ([TokenOrLexeme::Token(Token::GreaterThan)], Some('=')) => {
-            LexOperation::ShiftReduce(2, token!(Token::GreaterEqual))
-        }
-        ([lexeme!('>')], _) => LexOperation::Reduce(1, token!(Token::GreaterThan)),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::GreaterThan,
+                ..
+            })],
+            Some((end, '>')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::RightShift
+            )),
+        ),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::RightShift,
+                ..
+            })],
+            Some((end, '=')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::RightShiftEqual
+            )),
+        ),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::GreaterThan,
+                ..
+            })],
+            Some((end, '=')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::GreaterEqual
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '>')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::GreaterThan
+            )),
+        ),
 
-        ([TokenOrLexeme::Token(Token::LessThan)], Some('<')) => {
-            LexOperation::ShiftReduce(2, token!(Token::LShift))
-        }
-        ([TokenOrLexeme::Token(Token::LessThan)], Some('=')) => {
-            LexOperation::ShiftReduce(2, token!(Token::LessEqual))
-        }
-        ([lexeme!('<')], _) => LexOperation::Reduce(1, token!(Token::LessThan)),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::LessThan,
+                ..
+            })],
+            Some((end, '<')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::LeftShift
+            )),
+        ),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::LeftShift,
+                ..
+            })],
+            Some((end, '=')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::LeftShiftEqual
+            )),
+        ),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::LessThan,
+                ..
+            })],
+            Some((end, '=')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::LessEqual
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '<')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::LessThan
+            )),
+        ),
 
-        ([TokenOrLexeme::Token(Token::Ampersand)], Some('&')) => {
-            LexOperation::ShiftReduce(2, token!(Token::AmpersandAmperand))
-        }
-        ([lexeme!('&')], _) => LexOperation::Reduce(1, token!(Token::Ampersand)),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::Ampersand,
+                ..
+            })],
+            Some((end, '&')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::AmpersandAmpersand
+            )),
+        ),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::Ampersand,
+                ..
+            })],
+            Some((end, '=')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::AmpersandEqual
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '&')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::Ampersand
+            )),
+        ),
 
-        ([TokenOrLexeme::Token(Token::Pipe)], Some('|')) => {
-            LexOperation::ShiftReduce(2, token!(Token::PipePipe))
-        }
-        ([lexeme!('|')], _) => LexOperation::Reduce(1, token!(Token::Pipe)),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::Pipe,
+                ..
+            })],
+            Some((end, '|')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::PipePipe
+            )),
+        ),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::Pipe,
+                ..
+            })],
+            Some((end, '=')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::PipeEqual
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '|')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::Pipe
+            )),
+        ),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::Caret,
+                ..
+            })],
+            Some((end, '=')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::CaretEqual
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '^')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::Caret
+            )),
+        ),
 
-        ([lexeme!('^')], _) => LexOperation::Reduce(1, token!(Token::Carat)),
-        ([lexeme!('~')], _) => LexOperation::Reduce(1, token!(Token::Tilde)),
+        ([TokenOrLexeme::Lexeme(cur, '~')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::Tilde
+            )),
+        ),
 
-        ([TokenOrLexeme::Token(Token::Plus)], Some('+')) => {
-            LexOperation::ShiftReduce(2, token!(Token::PlusPlus))
-        }
-        ([lexeme!('+')], _) => LexOperation::Reduce(1, token!(Token::Plus)),
-        ([TokenOrLexeme::Token(Token::Minus)], Some('-')) => {
-            LexOperation::ShiftReduce(2, token!(Token::MinusMinus))
-        }
-        ([lexeme!('-')], _) => LexOperation::Reduce(1, token!(Token::Minus)),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::Plus,
+                ..
+            })],
+            Some((end, '+')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::PlusPlus
+            )),
+        ),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::Plus,
+                ..
+            })],
+            Some((end, '=')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::PlusEqual
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '+')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::Plus
+            )),
+        ),
 
-        ([lexeme!('*')], _) => LexOperation::Reduce(1, token!(Token::Star)),
-        ([lexeme!('/')], _) => LexOperation::Reduce(1, token!(Token::Slash)),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::Minus,
+                ..
+            })],
+            Some((end, '-')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::MinusMinus
+            )),
+        ),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::Minus,
+                ..
+            })],
+            Some((end, '=')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::MinusEqual
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '-')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::Minus
+            )),
+        ),
 
-        ([lexeme!('\''), TokenOrLexeme::Lexeme(c), lexeme!('\'')], _) if c.is_ascii() => {
-            LexOperation::Reduce(3, token!(Token::CharLiteral(*c)))
-        }
-        ([lexeme!('\'')], Some(l)) if l.is_ascii() || l == '\\' => LexOperation::Shift(lexeme!(l)),
-        ([lexeme!('\''), TokenOrLexeme::Lexeme('\\'), TokenOrLexeme::Lexeme(c)], Some('\''))
-            if c.is_ascii() =>
-        {
-            let escaped = to_ascii_escaped_char(*c).unwrap();
-            LexOperation::ShiftReduce(4, token!(Token::CharLiteral(escaped)))
-        }
-        ([lexeme!('\''), TokenOrLexeme::Lexeme(c)], Some('\'')) if c.is_ascii() => {
-            LexOperation::ShiftReduce(3, token!(Token::CharLiteral(*c)))
-        }
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::Star,
+                ..
+            })],
+            Some((end, '=')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::StarEqual
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '*')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::Star
+            )),
+        ),
 
-        ([lexeme!('\"'), .., lexeme!('\\')], Some('\"')) => LexOperation::Shift(lexeme!('\"')),
-        ([lexeme!('\"'), .., lexeme!('\\')], Some(c)) => {
-            LexOperation::ShiftReduce(1, lexeme!(to_ascii_escaped_char(c).unwrap()))
-        }
-        ([lexeme!('\"'), inner @ ..], Some('\"')) => {
-            let inner_str = inner
-                .iter()
-                .flat_map(|tol| match tol {
-                    TokenOrLexeme::Token(_) => None,
-                    TokenOrLexeme::Lexeme(c) => Some(*c),
-                })
-                .collect();
-            let token_cnt = inner.len() + 2;
-            LexOperation::ShiftReduce(token_cnt, token!(Token::StrLiteral(inner_str)))
-        }
-        ([lexeme!('\"'), ..], Some(c)) => LexOperation::Shift(lexeme!(c)),
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::Slash,
+                ..
+            })],
+            Some((end, '=')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::SlashEqual
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '/')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::Slash
+            )),
+        ),
 
-        ([TokenOrLexeme::Lexeme(l)], _) if l.is_digit(10) => l
-            .to_digit(10)
-            .map(|d| d as isize)
-            .map(|lit| LexOperation::Reduce(1, token!(Token::IntLiteral(lit))))
-            .unwrap_or_else(|| LexOperation::Err(format!("Unable to convert {} to digit", l))),
-        ([TokenOrLexeme::Token(Token::IntLiteral(_))], Some(l)) if l.is_digit(10) => {
-            LexOperation::Shift(lexeme!(l))
-        }
-        ([TokenOrLexeme::Token(Token::IntLiteral(base_int_lit)), TokenOrLexeme::Lexeme(l)], _)
-            if l.is_digit(10) =>
-        {
-            l.to_digit(10)
-                .map(|d| d as isize)
-                .map(|next| {
-                    // This match guarantees there is atleast a base.
-                    // if that base is < 10 we need to scale it.
-                    let unadjusted_multiplier = (base_int_lit / 10) as u32;
-                    let multiplier = if unadjusted_multiplier == 0 {
-                        1
-                    } else {
-                        unadjusted_multiplier
-                    };
+        (
+            [TokenOrLexeme::Token(Token {
+                span: start,
+                kind: TokenKind::PercentSign,
+                ..
+            })],
+            Some((end, '=')),
+        ) => LexOperation::ShiftReduce(
+            2,
+            token!(Token::new(
+                Span::new(start.start, end.increment_column()),
+                TokenKind::PercentEqual
+            )),
+        ),
+        ([TokenOrLexeme::Lexeme(cur, '%')], _) => LexOperation::Reduce(
+            1,
+            token!(Token::new(
+                Span::new(*cur, cur.increment_column()),
+                TokenKind::PercentSign
+            )),
+        ),
 
-                    let scaled_base = base_int_lit * 10isize.pow(multiplier);
-                    let new_lit = scaled_base + next;
+        (
+            [TokenOrLexeme::Lexeme(start, '\''), TokenOrLexeme::Lexeme(_, c), TokenOrLexeme::Lexeme(end, '\'')],
+            _,
+        ) if c.is_ascii() => {
+            let start_idx = start.index + 1;
+            let end_idx = end.index;
+            let data = &source[start_idx..end_idx];
 
-                    LexOperation::Reduce(2, token!(Token::IntLiteral(new_lit)))
-                })
-                .unwrap_or_else(|| LexOperation::Err(format!("Unable to convert {} to digit", l)))
-        }
-
-        // Identifiers
-        ([contents @ ..], Some(l))
-            if matches_rules(contents, lexeme_is_identifier_ascii())
-                && matches_rules(&[lexeme!(l)], lexeme_is_identifier_ascii()) =>
-        {
-            LexOperation::Shift(lexeme!(l))
-        }
-        // if lookahead doesn't match an identifier reduce
-        ([contents @ ..], Some(l))
-            if matches_rules(contents, lexeme_is_identifier_ascii())
-                && !matches_rules(&[lexeme!(l)], lexeme_is_identifier_ascii()) =>
-        {
             LexOperation::Reduce(
-                contents.len(),
-                token!(Token::Identifier(
-                    contents
-                        .iter()
-                        .flat_map(|tol| tol.to_lexeme())
-                        .collect::<String>()
+                3,
+                token!(Token::with_data(
+                    Span::new(*start, end.increment_column()),
+                    data,
+                    TokenKind::CharacterConstant
                 )),
             )
         }
-        // if lookahead is empty, reduce.
-        ([contents @ ..], None) if matches_rules(contents, lexeme_is_identifier_ascii()) => {
+        ([TokenOrLexeme::Lexeme(_, '\'')], Some((cur, l))) if l.is_ascii() || l == '\\' => {
+            LexOperation::Shift(lexeme!(cur, l))
+        }
+        (
+            [TokenOrLexeme::Lexeme(start, '\''), TokenOrLexeme::Lexeme(_, '\\'), TokenOrLexeme::Lexeme(_, c)],
+            Some((end, '\'')),
+        ) if c.is_ascii() => {
+            let start_idx = start.index + 1;
+            let end_idx = end.index;
+            let data = &source[start_idx..end_idx];
+
+            LexOperation::ShiftReduce(
+                4,
+                token!(Token::with_data(
+                    Span::new(*start, end.increment_column()),
+                    data,
+                    TokenKind::CharacterConstant
+                )),
+            )
+        }
+        ([TokenOrLexeme::Lexeme(start, '\''), TokenOrLexeme::Lexeme(_, c)], Some((end, '\'')))
+            if c.is_ascii() =>
+        {
+            let start_idx = start.index + 1;
+            let end_idx = end.index;
+            let data = &source[start_idx..end_idx];
+
+            LexOperation::ShiftReduce(
+                3,
+                token!(Token::with_data(
+                    Span::new(*start, end.increment_column()),
+                    data,
+                    TokenKind::CharacterConstant
+                )),
+            )
+        }
+
+        ([TokenOrLexeme::Lexeme(_, '\"'), TokenOrLexeme::Lexeme(_, '\\')], Some((cur, '\"'))) => {
+            LexOperation::Shift(lexeme!(cur, '\"'))
+        }
+        ([TokenOrLexeme::Lexeme(_, '\"'), TokenOrLexeme::Lexeme(cur, '\\')], Some((_, c))) => {
+            LexOperation::ShiftReduce(2, lexeme!(*cur, to_ascii_escaped_char(c).unwrap()))
+        }
+        ([TokenOrLexeme::Lexeme(start, '\"'), inner @ ..], Some((end, '\"'))) => {
+            let token_cnt = inner.len() + 2;
+
+            let start_idx = start.index + 1;
+            let end_idx = end.index;
+            let data = &source[start_idx..end_idx];
+
+            LexOperation::ShiftReduce(
+                token_cnt,
+                token!(Token::with_data(
+                    Span::new(*start, end.increment_column()),
+                    data,
+                    TokenKind::StringLiteral
+                )),
+            )
+        }
+        ([TokenOrLexeme::Lexeme(_, '\"'), ..], Some((cur, c))) => {
+            LexOperation::Shift(lexeme!(cur, c))
+        }
+
+        ([TokenOrLexeme::Lexeme(cur, c)], _) if c.is_digit(10) => {
+            let start = cur.index;
+            let end = start + 1;
+            let data = &source[start..end];
+
             LexOperation::Reduce(
-                contents.len(),
-                token!(Token::Identifier(
-                    contents
-                        .iter()
-                        .flat_map(|tol| tol.to_lexeme())
-                        .collect::<String>()
+                1,
+                token!(Token::with_data(
+                    Span::new(*cur, cur.increment_column()),
+                    data,
+                    TokenKind::IntegerConstant
+                )),
+            )
+        }
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::IntegerConstant,
+                ..
+            })],
+            Some((cur, c)),
+        ) if c.is_digit(10) => {
+            let start = span.start.index;
+            let end = cur.index + 1;
+            let data = &source[start..end];
+
+            LexOperation::ShiftReduce(
+                2,
+                token!(Token::with_data(
+                    Span::new(span.start, cur.increment_column()),
+                    data,
+                    TokenKind::IntegerConstant,
+                )),
+            )
+        }
+
+        // Identifiers
+        ([contents @ ..], Some((cur, c)))
+            if matches_rules(contents, lexeme_is_identifier_ascii())
+                && matches_rules(&[lexeme!(cur, c)], lexeme_is_identifier_ascii()) =>
+        {
+            LexOperation::Shift(lexeme!(cur, c))
+        }
+
+        // if lookahead doesn't match an identifier reduce
+        ([TokenOrLexeme::Lexeme(start, start_c), contents @ ..], Some((cur, l)))
+            if matches_rules(contents, lexeme_is_identifier_ascii())
+                && matches_rules(
+                    &[TokenOrLexeme::Lexeme(*start, *start_c)],
+                    lexeme_is_identifier_ascii(),
+                )
+                && !matches_rules(&[lexeme!(cur, l)], lexeme_is_identifier_ascii()) =>
+        {
+            let token_cnt = contents.len() + 1;
+            // lookahead's cursor is the non-inclusive end.
+            let end = cur;
+            let data = &source[start.index..end.index];
+
+            LexOperation::Reduce(
+                token_cnt,
+                token!(Token::with_data(
+                    Span::new(*start, end),
+                    data,
+                    TokenKind::Identifier
+                )),
+            )
+        }
+        // end of input
+        (
+            [TokenOrLexeme::Lexeme(start, start_c), contents @ .., TokenOrLexeme::Lexeme(end, end_c)],
+            None,
+        ) if matches_rules(contents, lexeme_is_identifier_ascii())
+            && matches_rules(
+                &[TokenOrLexeme::Lexeme(*start, *start_c)],
+                lexeme_is_identifier_ascii(),
+            )
+            && matches_rules(
+                &[TokenOrLexeme::Lexeme(*end, *end_c)],
+                lexeme_is_identifier_ascii(),
+            ) =>
+        {
+            // increment column to extend past the end of input.
+            let end = end.increment_column();
+            let data = &source[start.index..end.index];
+            let token_cnt = contents.len() + 2;
+            // lookahead's cursor is the non-inclusive end.
+
+            LexOperation::Reduce(
+                token_cnt,
+                token!(Token::with_data(
+                    Span::new(*start, end),
+                    data,
+                    TokenKind::Identifier
                 )),
             )
         }
 
         // Type Keywords
-        ([TokenOrLexeme::Token(Token::Identifier(id))], _) if id == "signed" => {
-            LexOperation::Reduce(1, token!(Token::Signed))
-        }
-        ([TokenOrLexeme::Token(Token::Identifier(id))], _) if id == "unsigned" => {
-            LexOperation::Reduce(1, token!(Token::Unsigned))
-        }
-        ([TokenOrLexeme::Token(Token::Identifier(id))], _) if id == "void" => {
-            LexOperation::Reduce(1, token!(Token::Void))
-        }
-        ([TokenOrLexeme::Token(Token::Identifier(id))], _) if id == "char" => {
-            LexOperation::Reduce(1, token!(Token::Char))
-        }
-        ([TokenOrLexeme::Token(Token::Identifier(id))], _) if id == "short" => {
-            LexOperation::Reduce(1, token!(Token::Short))
-        }
-        ([TokenOrLexeme::Token(Token::Identifier(id))], _) if id == "int" => {
-            LexOperation::Reduce(1, token!(Token::Int))
-        }
-        ([TokenOrLexeme::Token(Token::Identifier(id))], _) if id == "long" => {
-            LexOperation::Reduce(1, token!(Token::Long))
-        }
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("auto"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Auto))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("break"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Break))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("case"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Case))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("char"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Char))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("const"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Const))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("continue"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Continue))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("default"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Default))),
 
-        // Other keywords
-        ([TokenOrLexeme::Token(Token::Identifier(id))], _) if id == "if" => {
-            LexOperation::Reduce(1, token!(Token::If))
-        }
-        ([TokenOrLexeme::Token(Token::Identifier(id))], _) if id == "else" => {
-            LexOperation::Reduce(1, token!(Token::Else))
-        }
-        ([TokenOrLexeme::Token(Token::Identifier(id))], _) if id == "for" => {
-            LexOperation::Reduce(1, token!(Token::For))
-        }
-        ([TokenOrLexeme::Token(Token::Identifier(id))], _) if id == "while" => {
-            LexOperation::Reduce(1, token!(Token::While))
-        }
-
-        ([TokenOrLexeme::Token(Token::Identifier(id))], _) if id == "return" => {
-            LexOperation::Reduce(1, token!(Token::Return))
-        }
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("do"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Do))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("double"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Double))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("else"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Else))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("enum"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Enum))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("extern"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Extern))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("float"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Float))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("for"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::For))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("goto"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Goto))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("if"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::If))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("int"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Int))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("long"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Long))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("register"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Register))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("return"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Return))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("short"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Short))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("signed"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Signed))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("sizeof"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::SizeOf))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("static"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Static))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("struct"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Struct))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("switch"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Switch))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("typedef"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::TypeDef))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("union"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Union))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("unsigned"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Unsigned))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("void"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Void))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("volatile"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::Volatile))),
+        (
+            [TokenOrLexeme::Token(Token {
+                span,
+                kind: TokenKind::Identifier,
+                data: Some("while"),
+            })],
+            _,
+        ) => LexOperation::Reduce(1, token!(Token::new(*span, TokenKind::While))),
 
         ([TokenOrLexeme::Token(_)], _) | ([.., TokenOrLexeme::Token(_)], _) => {
             LexOperation::NoFurtherRefinement
         }
-        ([.., TokenOrLexeme::Lexeme(c)], _) if c.is_whitespace() => match *c {
-            '\n' => LexOperation::Reduce(1, token!(Token::Newline)),
-            _ => LexOperation::Reduce(1, token!(Token::Space)),
+        ([.., TokenOrLexeme::Lexeme(_, c)], _) if c.is_whitespace() => match *c {
+            '\n' => LexOperation::Advance(1, AdvanceOperation::IncrementLine),
+            _ => LexOperation::Advance(1, AdvanceOperation::IncrementColumn),
         },
 
         _ => LexOperation::Err("No rules matched with unpaired lexemes.".to_string()),
     }
 }
 
+enum ScanState<T, E> {
+    Ok(T),
+    Err(E),
+    Done,
+}
+
 /// Scanner takes a string representing lox source and attempts to convert the
 /// source into a vector of either Tokens or lexical errors.
 pub struct Scanner<'a> {
-    source: core::iter::Peekable<core::str::Chars<'a>>,
+    source: &'a str,
+    iter: core::iter::Peekable<core::str::Chars<'a>>,
     cursor: Cursor,
-    stack: LexerStack<STACK_SIZE>,
+    stack: LexerStack<'a, STACK_SIZE>,
     end: usize,
 }
 
@@ -556,7 +1401,8 @@ impl<'a> Scanner<'a> {
         let stack = LexerStack::new([TOKEN_OR_LEXEME_INITIALIZER; STACK_SIZE]);
         let end = source.len();
         Scanner {
-            source: source.chars().peekable(),
+            source,
+            iter: source.chars().peekable(),
             cursor: Cursor::default(),
             stack,
             end,
@@ -568,67 +1414,57 @@ impl<'a> Scanner<'a> {
         index >= self.end
     }
 
-    fn scan_token_from(&mut self) -> Result<Token, String> {
-        if self.source.peek().is_none() {
-            Ok(Token::Eof)
+    fn scan_token_from(&mut self) -> ScanState<Token<'a>, String> {
+        use ScanState::*;
+
+        if self.iter.peek().is_none() {
+            ScanState::Done
         } else {
-            let mut lookahead = self.source.peek().copied();
+            let mut lookahead = self.iter.peek().copied();
             loop {
                 let stack = self.stack.as_ref();
-                let op = stack_eval_to_token(stack, lookahead);
+                let postional_char = lookahead.map(|c| (self.cursor, c));
+                let op = stack_eval_to_token(self.source, stack, postional_char);
 
                 match op {
-                    LexOperation::Shift(TokenOrLexeme::Lexeme('\n')) => {
-                        // safe to unwrap due to is_some guarantee
-                        self.stack.push_mut(lexeme!('\n'));
-                        lookahead = consume_and_peek_from_iter(&mut self.source);
-                        self.cursor.increment_line_mut();
+                    LexOperation::Shift(TokenOrLexeme::Lexeme(cur, '\n')) => {
+                        self.stack.push_mut(lexeme!(cur, '\n'));
+                        lookahead = consume_and_peek_from_iter(&mut self.iter);
                     }
                     LexOperation::Shift(next_val) => {
-                        // safe to unwrap due to is_some guarantee
                         self.stack.push_mut(next_val);
-                        lookahead = consume_and_peek_from_iter(&mut self.source);
-                        self.cursor.increment_column_mut();
-                    }
-                    LexOperation::ShiftReduce(by, TokenOrLexeme::Token(Token::Newline)) => {
-                        for _ in 0..by {
-                            self.stack.pop_mut();
-                        }
-                        self.cursor.increment_line_mut();
-                    }
-                    LexOperation::ShiftReduce(by, TokenOrLexeme::Token(Token::Space)) => {
-                        for _ in 0..by {
-                            self.stack.pop_mut();
-                        }
+                        lookahead = consume_and_peek_from_iter(&mut self.iter);
                         self.cursor.increment_column_mut();
                     }
                     LexOperation::ShiftReduce(by, TokenOrLexeme::Token(to_token)) => {
-                        lookahead = consume_and_peek_from_iter(&mut self.source);
+                        lookahead = consume_and_peek_from_iter(&mut self.iter);
                         // pop n elements off the stack
                         for _ in 0..by {
                             self.stack.pop_mut();
                         }
 
                         self.stack.push_mut(TokenOrLexeme::Token(to_token));
-                        // safe to unwrap due to is_some guarantee
                         self.cursor.increment_column_mut();
                     }
 
-                    LexOperation::ShiftReduce(by, TokenOrLexeme::Lexeme(next_val)) => {
-                        lookahead = consume_and_peek_from_iter(&mut self.source);
+                    LexOperation::ShiftReduce(by, TokenOrLexeme::Lexeme(cur, next_val)) => {
+                        lookahead = consume_and_peek_from_iter(&mut self.iter);
                         for _ in 0..by {
                             self.stack.pop_mut();
                         }
-                        self.stack.push_mut(lexeme!(next_val));
+                        self.stack.push_mut(lexeme!(cur, next_val));
                         self.cursor.increment_column_mut();
                     }
-                    LexOperation::Reduce(by, TokenOrLexeme::Token(Token::Newline)) => {
+                    LexOperation::Advance(by, AdvanceOperation::IncrementLine) => {
+                        // pop the newline off the stack.
                         for _ in 0..by {
                             self.stack.pop_mut();
                         }
+
+                        self.cursor.increment_line_mut();
                     }
                     // this will only occur with whitespace
-                    LexOperation::Reduce(by, TokenOrLexeme::Token(Token::Space)) => {
+                    LexOperation::Advance(by, AdvanceOperation::IncrementColumn) => {
                         for _ in 0..by {
                             self.stack.pop_mut();
                         }
@@ -644,11 +1480,13 @@ impl<'a> Scanner<'a> {
                         self.stack.push_mut(tol);
                     }
                     LexOperation::NoFurtherRefinement => match self.stack.pop_mut() {
-                        Some(TokenOrLexeme::Token(tok)) => return Ok(tok),
+                        Some(TokenOrLexeme::Token(tok)) => return ScanState::Ok(tok),
                         _ => return Err(format!("invalid token at {}", &self.cursor)),
                     },
 
-                    LexOperation::Err(e) => return Err(format!("{} at {}", e, &self.cursor)),
+                    LexOperation::Err(e) => {
+                        return ScanState::Err(format!("{} at {}", e, &self.cursor))
+                    }
                 }
             }
         }
@@ -656,7 +1494,7 @@ impl<'a> Scanner<'a> {
 }
 
 impl<'a> IntoIterator for Scanner<'a> {
-    type Item = Result<Token, String>;
+    type Item = Result<Token<'a>, String>;
     type IntoIter = ScannerIntoIterator<'a>;
 
     fn into_iter(self) -> Self::IntoIter {
@@ -669,7 +1507,7 @@ pub struct ScannerIntoIterator<'a> {
 }
 
 impl<'a> Iterator for ScannerIntoIterator<'a> {
-    type Item = Result<Token, String>;
+    type Item = Result<Token<'a>, String>;
 
     fn next(&mut self) -> Option<Self::Item> {
         if self.scanner.is_empty() {
@@ -677,9 +1515,9 @@ impl<'a> Iterator for ScannerIntoIterator<'a> {
         } else {
             let next = self.scanner.scan_token_from();
             match next {
-                Ok(Token::Eof) => None,
-                lex_res @ Ok(_) => Some(lex_res),
-                lex_res @ Err(_) => Some(lex_res),
+                ScanState::Done => None,
+                ScanState::Ok(token) => Some(Ok(token)),
+                ScanState::Err(e) => Some(Err(e)),
             }
         }
     }
@@ -690,11 +1528,12 @@ impl<'a> Iterator for ScannerIntoIterator<'a> {
 ///
 /// # Notes
 /// This requires a needless collect to prevent a stack error.
-#[allow(clippy::needless_collect)]
+#[allow(clippy::needless_collect, unused)]
 pub fn lex(src: &str) -> LexResult {
     let scanner = Scanner::new(src);
     let tokens: Vec<Result<Token, LexError>> = scanner.into_iter().collect();
     let collected_result: Result<Vec<Token>, String> = tokens.into_iter().collect();
+
     collected_result
 }
 
@@ -703,28 +1542,59 @@ mod tests {
     use super::*;
 
     #[test]
-    fn should_scan_single_token() {
-        let inputs = ["{", " {", " { "];
+    fn should_generate_sub_str_from_token() {
+        let input = "unsigned int";
 
-        for input in inputs {
+        let span = Span::new(Cursor::new(9, 9, 0), Cursor::new(12, 12, 0));
+        let token = Token::new(span, TokenKind::Int);
+
+        assert_eq!(Some("int"), token.str_from_input(input))
+    }
+
+    #[test]
+    fn should_scan_single_token() {
+        let inputs = [(0, "{"), (1, " {"), (1, " { ")];
+
+        for (pos, input) in inputs.iter() {
             let mut scanner = Scanner::new(input).into_iter();
 
-            assert_eq!(Some(Ok(Token::LBrace)), scanner.next())
+            let start = *pos;
+            let end = start + 1;
+
+            assert_eq!(
+                Some(Ok(Token::new(
+                    Span::new(Cursor::new(start, start, 1), Cursor::new(end, end, 1)),
+                    TokenKind::LeftBrace
+                ))),
+                scanner.next()
+            )
         }
     }
 
     #[test]
     fn should_scan_compound_token() {
         let input_expected = [
-            (["<", " <", " < "], Token::LessThan),
-            (["<=", " <=", " <= "], Token::LessEqual),
+            (
+                [(0, 1, "<"), (1, 2, " <"), (1, 2, " < ")],
+                TokenKind::LessThan,
+            ),
+            (
+                [(0, 2, "<="), (1, 3, " <="), (1, 3, " <= ")],
+                TokenKind::LessEqual,
+            ),
         ];
 
-        for (inputs, expected) in input_expected {
-            for input in inputs {
+        for (inputs, expected_kind) in input_expected {
+            for (start, end, input) in inputs {
                 let mut scanner = Scanner::new(input).into_iter();
 
-                assert_eq!(Some(Ok(expected.clone())), scanner.next())
+                assert_eq!(
+                    Some(Ok(Token::new(
+                        Span::new(Cursor::new(start, start, 1), Cursor::new(end, end, 1)),
+                        expected_kind
+                    ))),
+                    scanner.next()
+                )
             }
         }
     }
@@ -732,88 +1602,110 @@ mod tests {
     #[test]
     fn should_scan_digit_token() {
         let input_expected = [
-            (["1", " 1", " 1 "], Token::IntLiteral(1)),
-            (["123", " 123", " 123 "], Token::IntLiteral(123)),
+            ([(0, 1, "1"), (1, 2, " 1"), (1, 2, " 1 ")], "1"),
+            ([(0, 3, "123"), (1, 4, " 123"), (1, 4, " 123 ")], "123"),
         ];
 
-        for (inputs, expected) in input_expected {
-            for input in inputs {
+        for (inputs, data) in input_expected {
+            for (start, end, input) in inputs {
                 let mut scanner = Scanner::new(input).into_iter();
 
-                assert_eq!(Some(Ok(expected.clone())), scanner.next())
+                assert_eq!(
+                    Some(Ok(Token::with_data(
+                        Span::new(Cursor::new(start, start, 1), Cursor::new(end, end, 1)),
+                        data,
+                        TokenKind::IntegerConstant
+                    ))),
+                    scanner.next()
+                )
             }
         }
     }
 
     #[test]
     fn should_scan_char_literals() {
-        let input_expected = [
-            (["\'a\'"], Token::CharLiteral('a')),
-            ([" \'a\'"], Token::CharLiteral('a')),
-            ([" \'a\' "], Token::CharLiteral('a')),
-            (["\'\t\'"], Token::CharLiteral('\t')),
-            ([" \'\t\'"], Token::CharLiteral('\t')),
-            ([" \'\t\' "], Token::CharLiteral('\t')),
+        let inputs = [
+            (0, 3, "\'a\'", "a"),
+            (1, 4, " \'a\'", "a"),
+            (1, 4, " \'a\' ", "a"),
+            (0, 3, "\'\t\'", "\t"),
+            (1, 4, " \'\t\'", "\t"),
+            (0, 3, "\'\t\' ", "\t"),
         ];
 
-        for (inputs, expected) in input_expected {
-            for input in inputs {
-                let mut scanner = Scanner::new(input).into_iter();
+        for (start, end, input, data) in inputs {
+            let mut scanner = Scanner::new(input).into_iter();
 
-                assert_eq!(Some(Ok(expected.clone())), scanner.next())
-            }
+            assert_eq!(
+                Some(Ok(Token::with_data(
+                    Span::new(Cursor::new(start, start, 1), Cursor::new(end, end, 1)),
+                    data,
+                    TokenKind::CharacterConstant
+                ))),
+                scanner.next()
+            )
         }
     }
 
     #[test]
     fn should_scan_string_literals() {
         let input_expected = [
-            (["\"hello\""], Token::StrLiteral("hello".to_string())),
-            ([" \"hello\""], Token::StrLiteral("hello".to_string())),
-            ([" \"hello\" "], Token::StrLiteral("hello".to_string())),
-            (["\"hello\\n\""], Token::StrLiteral("hello\n".to_string())),
+            (0, 7, "\"hello\"", "hello"),
+            (1, 8, " \"hello\"", "hello"),
+            (1, 8, " \"hello\" ", "hello"),
+            (0, 9, "\"hello\\n\"", "hello\\n"),
         ];
 
-        for (inputs, expected) in input_expected {
-            for input in inputs {
-                let mut scanner = Scanner::new(input).into_iter();
+        for (start, end, input, expected_data) in input_expected {
+            let mut scanner = Scanner::new(input).into_iter();
 
-                assert_eq!(Some(Ok(expected.clone())), scanner.next())
-            }
+            assert_eq!(
+                Some(Ok(Token::with_data(
+                    Span::new(Cursor::new(start, start, 1), Cursor::new(end, end, 1)),
+                    expected_data,
+                    TokenKind::StringLiteral
+                ))),
+                scanner.next()
+            )
         }
     }
 
     #[test]
     fn should_scan_identifier() {
         let input_expected = [
-            (["test"], Token::Identifier("test".to_string())),
-            ([" test"], Token::Identifier("test".to_string())),
-            ([" test "], Token::Identifier("test".to_string())),
+            (0, 4, "test", "test"),
+            (1, 5, " test", "test"),
+            (1, 5, " test ", "test"),
         ];
 
-        for (inputs, expected) in input_expected {
-            for input in inputs {
-                let mut scanner = Scanner::new(input).into_iter();
+        for (start, end, input, expected_data) in input_expected {
+            let mut scanner = Scanner::new(input).into_iter();
 
-                assert_eq!(Some(Ok(expected.clone())), scanner.next())
-            }
+            assert_eq!(
+                Some(Ok(Token::with_data(
+                    Span::new(Cursor::new(start, start, 1), Cursor::new(end, end, 1)),
+                    expected_data,
+                    TokenKind::Identifier
+                ))),
+                scanner.next()
+            )
         }
     }
 
     #[test]
     fn should_scan_keywords() {
-        let input_expected = [
-            (["for"], Token::For),
-            ([" for"], Token::For),
-            ([" for "], Token::For),
-        ];
+        let input_expected = [(0, 3, "for"), (1, 4, " for"), (1, 4, " for ")];
 
-        for (inputs, expected) in input_expected {
-            for input in inputs {
-                let mut scanner = Scanner::new(input).into_iter();
+        for (start, end, input) in input_expected {
+            let mut scanner = Scanner::new(input).into_iter();
 
-                assert_eq!(Some(Ok(expected.clone())), scanner.next())
-            }
+            assert_eq!(
+                Some(Ok(Token::new(
+                    Span::new(Cursor::new(start, start, 1), Cursor::new(end, end, 1)),
+                    TokenKind::For
+                ))),
+                scanner.next()
+            )
         }
     }
 
@@ -823,14 +1715,44 @@ mod tests {
         let res = lex(input);
 
         let expected: Result<Vec<_>, _> = Ok(vec![
-            Token::LBrace,
-            Token::IntLiteral(1),
-            Token::Plus,
-            Token::IntLiteral(2),
-            Token::SemiColon,
-            Token::RBrace,
+            Token::new(
+                Span::new(Cursor::new(0, 0, 1), Cursor::new(1, 1, 1)),
+                TokenKind::LeftBrace,
+            ),
+            Token::with_data(
+                Span::new(Cursor::new(2, 2, 1), Cursor::new(3, 3, 1)),
+                "1",
+                TokenKind::IntegerConstant,
+            ),
+            Token::new(
+                Span::new(Cursor::new(4, 4, 1), Cursor::new(5, 5, 1)),
+                TokenKind::Plus,
+            ),
+            Token::with_data(
+                Span::new(Cursor::new(6, 6, 1), Cursor::new(7, 7, 1)),
+                "2",
+                TokenKind::IntegerConstant,
+            ),
+            Token::new(
+                Span::new(Cursor::new(7, 7, 1), Cursor::new(8, 8, 1)),
+                TokenKind::SemiColon,
+            ),
+            Token::new(
+                Span::new(Cursor::new(9, 9, 1), Cursor::new(10, 10, 1)),
+                TokenKind::RightBrace,
+            ),
         ]);
 
         assert_eq!(expected, res)
+    }
+
+    #[test]
+    fn should_format_span_as_expected() {
+        let span = Span::new(Cursor::new(9, 9, 0), Cursor::new(12, 12, 1));
+
+        assert_eq!(
+            "(line: 0, column: 9)..(line: 1, column: 12)",
+            format!("{}", span)
+        )
     }
 }
